+++ src/main/java/com/back/Main.java
@@ -1,0 +1,12 @@
+package com.back;
+
+import java.util.ArrayList;
+import java.util.Scanner;
+
+public class Main {
+    public static void main(String[] args) {
+        App app = new App();
+        app.start();
+    }
+
+}
